backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.3
pydantic-settings==2.6.1
redis==5.2.1
python-dotenv==1.0.1
earthaccess==0.12.0
xarray==2024.11.0
pandas==2.2.3
numpy==2.2.1
h5netcdf==1.4.1
netCDF4==1.7.2
scikit-learn==1.6.0
joblib==1.4.2
python-multipart==0.0.20
pytest==8.3.3
